Revert "Snapshot shutdown progress tracker test fix (#139447)"
This reverts commit 76e9a5a972fd9c001d30461bb4d12b618b243afd.
