+++ server/src/internalClusterTest/java/org/elasticsearch/repositories/SnapshotMetricsIT.java
@@ -45,8 +45,6 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import java.util.stream.StreamSupport;
 
-import static org.elasticsearch.snapshots.SnapshotTestUtils.flushMasterQueue;
-import static org.elasticsearch.snapshots.SnapshotTestUtils.putShutdownForRemovalMetadata;
 import static org.elasticsearch.threadpool.ThreadPool.ESTIMATED_TIME_INTERVAL_SETTING;
 import static org.hamcrest.Matchers.allOf;
 import static org.hamcrest.Matchers.empty;
@@ -298,7 +296,7 @@ public class SnapshotMetricsIT extends AbstractSnapshotIntegTestCase {
         assertMetricsHaveAttributes(InstrumentType.LONG_COUNTER, SnapshotMetrics.SNAPSHOT_CREATE_THROTTLE_DURATION, expectedAttrs);
     }
 
-    public void testByStateCounts_InitAndQueuedShards() {
+    public void testByStateCounts_InitAndQueuedShards() throws Exception {
         final String indexName = randomIdentifier();
         final int numShards = randomIntBetween(2, 10);
         final int numReplicas = randomIntBetween(0, 1);
@@ -438,7 +436,7 @@ public class SnapshotMetricsIT extends AbstractSnapshotIntegTestCase {
         );
     }
 
-    public void testByStateCounts_WaitingShards() {
+    public void testByStateCounts_WaitingShards() throws Exception {
         final String indexName = randomIdentifier();
         final String boundNode = internalCluster().startDataOnlyNode();
         final String destinationNode = internalCluster().startDataOnlyNode();
